🗃️ Committing everything that changed 🗃️

+++ root/.local/bin/update-resolv.sh
@@ -22,10 +22,8 @@ __fetch() {
   return $?
 }
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-if [ -f "/etc/resolv.conf " ]; then
-  chattr -i "/etc/resolv.conf"
-  mv -f "/etc/resolv.conf" "/tmp/resolv.bak.conf"
-fi
+[ -f "/etc/resolv.conf" ] && chattr -i "/etc/resolv.conf" || echo "nameserver 1.1.1.1" >"/etc/resolv.conf"
+[ -f "/etc/resolv.conf" ] && mv -f "/etc/resolv.conf" "/tmp/resolv.bak.conf"
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 if __fetch; then
   [ -f "/tmp/resolv.conf" ] && mv -f "/tmp/resolv.conf" "/etc/resolv.conf"
@@ -36,7 +34,7 @@ fi
 [ -f "/tmp/resolv.conf" ] && rm -Rf "/tmp/resolv.conf"
 [ -f "/tmp/resolv.bak.conf" ] && rm -Rf "/tmp/resolv.bak.conf"
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-chattr +i "/etc/resolv.conf"
+[ -f "/etc/resolv.conf" ] && chattr +i "/etc/resolv.conf"
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 exit
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
